LFS-241: Add support for questionnaire sections Move CONTAINER_PROPS to QuestionnaireStyle.jsx
