+++ modules/data-entry/src/main/frontend/src/questionnaire/Subject.jsx
@@ -29,7 +29,7 @@ import {
 
 import LiveTable from "../dataHomepage/LiveTable.jsx";
 
-import QuestionnaireStyle from "./QuestionnaireStyle";
+import QuestionnaireStyle, { CONTAINER_PROPS } from "./QuestionnaireStyle";
 
 /**
  * Component that displays a Subject.
@@ -120,7 +120,7 @@ function Subject (props) {
 
   return (
     <div>
-      <Grid container direction="column" spacing={4} alignItems="stretch" justify="space-between" wrap="nowrap">
+      <Grid container {...CONTAINER_PROPS} >
         <Grid item>
           {
             data && data.identifier ?
@@ -145,5 +145,9 @@ function Subject (props) {
     </div>
   );
 };
+
+Subject.PropTypes = {
+  id: PropTypes.string
+}
 
 export default withStyles(QuestionnaireStyle)(Subject);
